Add files via upload
Added filtering for AUX data.

Had som instances where I had 0 voltage and data in some of the column_N fields. Found that coulmn_1 was to identify if the following dataset is from AUX channel (temp and voltage) or not. Aux data has the same STEP_ID but data field begins with 1 and ends with 65 where cycler data begins with 0 and ends with 55