Add files via upload
I've added a function to the nda_version_8_0_beta.py file which replaces the old current_multiplier if statement. It works with my files but because my data uses quite a narrow current range, I've decided to place the functionality it in its own file for now.
You can use it by setting "beta=True" in read_nda() or nda_to_csv().
Once we're satified that this works, I'll replace the nda_version_8_0.py file with the beta code and remove the beta option.